Refactor: #46 - 파일 경로 수정
